requirements.txt
pyspark==3.5.0
pandas==2.0.3
requests==2.31.0
boto3==1.28.57
beautifulsoup4==4.12.2
lxml==4.9.3
dbt-core==1.6.0
dbt-postgres==1.6.0
psycopg2-binary==2.9.7
jupyter==1.0.0
pytest==7.4.0